re-structures the projects with packages
