+++ .pre-commit-config.yaml
@@ -56,6 +56,6 @@ repos:
   - id: flake8
     args:
       - "--max-line-length=120"
-      - "--per-file-ignores=__init__.py:F401"
+      - "--per-file-ignores=base.py:F401"
   ## You can add flake8 plugins via `additional_dependencies`:
   #  additional_dependencies: [flake8-bugbear]
