+++ src/dimcat/analyzer/__init__.py
@@ -1,0 +1,2 @@
+from .slices import LocalKeySequence, LocalKeyUnique
+from .facet import TPCrange, PitchClassVectors, ChordSymbolUnigrams, ChordSymbolBigrams
