+++ src/dimcat/analyzer/facet.py
@@ -1,152 +1,12 @@
-"""Analyzers are PipelineSteps that process data and store the results in Data.processed."""
-import logging
-from abc import ABC, abstractmethod
-from typing import (
-    Any,
-    Collection,
-    Iterable,
-    Iterator,
-    List,
-    Literal,
-    Tuple,
-    Type,
-    TypeVar,
-    Union,
-)
+from typing import Iterator, Tuple, Literal, Iterable
 
-import dimcat.data as data_module
 import pandas as pd
 from ms3 import add_weighted_grace_durations, fifths2name
 
-from ._typing import ID
-from .data import AnalyzedData, AnalyzedGroupedDataset, AnalyzedSlicedDataset, _Dataset
-from .pipeline import _STR2STEP, PipelineStep
-from .utils import grams, make_suffix
-
-logger = logging.getLogger(__name__)
-
-
-def _stepstrings2steps(stepstrings: Union[str, Collection[str]]) -> Tuple[Type, ...]:
-    """Turns one or several names of :obj:`PipelineStep` objects into a
-    tuple of references to these classes."""
-    if isinstance(stepstrings, str):
-        stepstrings = [stepstrings]
-    result = [_STR2STEP[step] for step in stepstrings]
-    return tuple(result)
-
-
-def _typestring2type(typestring: str) -> Type:
-    return getattr(data_module, typestring)
-
-
-R = TypeVar("R")
-
-
-class Analyzer(PipelineStep, ABC):
-    """Analyzers are PipelineSteps that process data and store the results in Data.processed."""
-
-    result_type = "Result"
-
-    assert_steps: Union[str, Collection[str]] = []
-    """Analyzer.process_data() raises ValueError if at least one of the names does not belong to
-    a :obj:`PipelineStep` that is among the previous PipelineSteps applied to the :obj:`_Dataset`."""
-
-    assert_previous_step: Union[str, Collection[str]] = []
-    """Analyzer.process_data() raises ValueError if last :obj:`PipelineStep` applied to the
-    :obj:`_Dataset` does not match any of these types."""
-
-    excluded_steps: Union[str, Collection[str]] = []
-    """Analyzer.process_data() raises ValueError if any of the previous :obj:`PipelineStep` applied to the
-    :obj:`_Dataset` matches one of these types."""
-
-    def __init__(self):
-        """Creates essential fields."""
-        self.config = {}
-        """:obj:`dict`
-        This dictionary stores the parameters to be passed to the compute() method."""
-        self.group2pandas = None
-        """:obj:`str`
-        The name of the function that allows displaying one group's results as a single
-        pandas object. See data.Corpus.convert_group2pandas()"""
-        self.level_names = {}
-        """:obj:`dict`
-        Define {"indices": "index_level_name"} if the analysis is applied once per group,
-        because the index of the DataFrame holding the processed data won't be showing the
-        individual indices anymore.
-        """
-
-    @staticmethod
-    @abstractmethod
-    def aggregate(result_a: R, result_b: R) -> R:
-        """Static method that combines two results of :meth:`compute`.
-
-        This needs to be equivalent to calling self.compute on the concatenation of the respective data resulting
-        in the two arguments."""
-        pass
-
-    @staticmethod
-    @abstractmethod
-    def compute(self, **kwargs) -> R:
-        """Static method that performs the actual computation takes place."""
-
-    @abstractmethod
-    def data_iterator(self, data: AnalyzedData) -> Iterator[Tuple[ID, Any]]:
-        """How a particular analyzer iterates through a dataset, getting the chunks passed to :meth:`compute`."""
-        yield from data
-
-    def process_data(self, dataset: _Dataset) -> AnalyzedData:
-        """Returns an :obj:`AnalyzedData` copy of the Dataset with the added analysis result."""
-        analyzer_name = self.__class__.__name__
-        if len(self.assert_steps) > 0:
-            assert_steps = _stepstrings2steps(self.assert_steps)
-            for step in assert_steps:
-                if not any(
-                    isinstance(previous_step, step)
-                    for previous_step in dataset.pipeline_steps
-                ):
-                    raise ValueError(
-                        f"{analyzer_name} require previous application of a {step.__name__}."
-                    )
-        if len(self.assert_previous_step) > 0:
-            assert_previous_step = _stepstrings2steps(self.assert_previous_step)
-            previous_step = dataset.pipeline_steps[0]
-            if not isinstance(previous_step, assert_previous_step):
-                raise ValueError(
-                    f"{analyzer_name} requires the previous pipeline step to be an "
-                    f"instance of {self.assert_previous_step}, not {previous_step.__name__}."
-                )
-        if len(self.excluded_steps) > 0:
-            excluded_steps = _stepstrings2steps(self.excluded_steps)
-            for step in excluded_steps:
-                if any(
-                    isinstance(previous_step, step)
-                    for previous_step in dataset.pipeline_steps
-                ):
-                    raise ValueError(
-                        f"{analyzer_name} cannot be applied when a {step.__name__} has been applied before."
-                    )
-        new_dataset = AnalyzedData(dataset)
-        result_type = _typestring2type(self.result_type)
-        result_object = result_type(
-            analyzer=self, dataset_before=dataset, dataset_after=new_dataset
-        )
-        result_object.config = self.config
-        for idx, df in self.data_iterator(new_dataset):
-            eligible, message = self.check(df)
-            if not eligible:
-                logger.info(f"{idx}: {message}")
-                continue
-            result_object[idx] = self.compute(df, **self.config)
-        result_object = self.post_process(result_object)
-        new_dataset.track_pipeline(
-            self, group2pandas=self.group2pandas, **self.level_names
-        )
-        new_dataset.set_result(self, result_object)
-        return new_dataset
-
-    def post_process(self, processed):
-        """Whatever needs to be done after analyzing the data before passing it to the dataset."""
-        return processed
+from dimcat._typing import ID
+from dimcat.analyzer.base import Analyzer
+from dimcat.data import AnalyzedData
+from dimcat.utils import make_suffix, grams
 
 
 class FacetAnalyzer(Analyzer):
@@ -467,76 +327,3 @@ class ChordSymbolBigrams(ChordSymbolAnalyzer):
             print(expanded)
             raise
         return counts
-
-
-class SliceInfoAnalyzer(Analyzer):
-    """"""
-
-    assert_steps = ["Slicer"]
-
-    def check(self, df):
-        if len(df) == 0:
-            return False, "DataFrame is empty."
-        return True, ""
-
-    def data_iterator(
-        self, data: AnalyzedSlicedDataset
-    ) -> Iterator[Tuple[Literal["all_slices"], pd.DataFrame]]:
-        yield from [("all_slices", data.get_slice_info())]
-
-
-class GroupedSliceInfoAnalyzer(SliceInfoAnalyzer):
-    """"""
-
-    assert_steps = ["Slicer", "Grouper"]
-
-    def data_iterator(
-        self, data: AnalyzedGroupedDataset
-    ) -> Iterator[Tuple[ID, pd.DataFrame]]:
-        yield from data.iter_grouped_slice_info()
-
-
-class LocalKeySequence(GroupedSliceInfoAnalyzer):
-    """"""
-
-    assert_steps = ["LocalKeySlicer", "PieceGrouper"]
-
-    def __init__(
-        self,
-    ):
-        """"""
-        super().__init__()
-        self.level_names["processed"] = ["localkeys"]
-        self.group2pandas = None
-
-    @staticmethod
-    def aggregate(result_a: pd.Series, result_b: pd.Series) -> pd.Series:
-        return pd.concat(result_a, result_b)
-
-    @staticmethod
-    def compute(slice_info: pd.DataFrame) -> pd.Series:
-        return slice_info.localkey
-
-
-class LocalKeyUnique(SliceInfoAnalyzer):
-    """"""
-
-    assert_steps = ["LocalKeySlicer"]
-
-    def __init__(
-        self,
-    ):
-        """"""
-        super().__init__()
-        self.level_names["processed"] = ["localkeys"]
-        self.group2pandas = None
-
-    @staticmethod
-    def aggregate(result_a: List[str], result_b: List[str]) -> List[str]:
-        set_a = set(result_a)
-        set_a.update(result_b)
-        return sorted(set_a)
-
-    @staticmethod
-    def compute(slice_info: pd.DataFrame) -> List[str]:
-        return list(slice_info.localkey.unique())
