+++ src/dimcat/data/__init__.py
@@ -1,0 +1,3 @@
+from .base import _Dataset, AnalyzedData, GroupedData, SlicedData
+from .dcml import Dataset, AnalyzedDataset, GroupedDataset, SlicedDataset, AnalyzedGroupedDataset, AnalyzedSlicedDataset, AnalyzedGroupedSlicedDataset, GroupedSlicedDataset
+from .result import Result
