+++ src/dimcat/grouper/__init__.py
@@ -1,0 +1,1 @@
+from .base import CorpusGrouper, ModeGrouper, PieceGrouper, YearGrouper
