+++ src/dimcat/grouper/base.py
@@ -3,10 +3,10 @@ import logging
 from abc import ABC, abstractmethod
 from collections import defaultdict
 
-from .data import AnalyzedData, GroupedData, _Dataset
-from .pipeline import PipelineStep
-from .slicer import LocalKeySlicer
-from .utils import get_composition_year
+from dimcat.data import AnalyzedData, GroupedData, _Dataset
+from dimcat.base import PipelineStep
+from dimcat.slicer import LocalKeySlicer
+from dimcat.utils import get_composition_year
 
 logger = logging.getLogger(__name__)
