+++ src/dimcat/slicer/__init__.py
@@ -1,0 +1,8 @@
+from .base import (
+    ChordCriterionSlicer,
+    ChordFeatureSlicer,
+    LocalKeySlicer,
+    MeasureSlicer,
+    NoteSlicer,
+    PhraseSlicer,
+)
