+++ src/dimcat/utils/__init__.py
@@ -1,0 +1,1 @@
+from .functions import *
