+++ src/dimcat/utils/functions.py
@@ -1,7 +1,8 @@
 """Utility functions that are or might be used by several modules or useful in external contexts."""
-from typing import Collection
+from typing import Collection, Union, Tuple, Type
 
 import pandas as pd
+from dimcat.base import Data, PipelineStep
 
 
 def nest_level(obj, include_tuples=False):
@@ -129,3 +130,23 @@ def interval_index2interval(ix):
     left = ix.left.min()
     right = ix.right.max()
     return pd.Interval(left, right, closed="left")
+
+
+
+def typestrings2types(typestrings: Union[str, Collection[str]]) -> Tuple[Type]:
+    """Turns one or several names of classes contained in this module into a
+    tuple of references to these classes."""
+    global Data, PipelineStep
+    d_types = Data._registry
+    ps_types = PipelineStep._registry
+    if isinstance(typestrings, str):
+        typestrings = [typestrings]
+    result = []
+    for typ in typestrings:
+        if typ in d_types:
+            result.append(d_types[typ])
+        elif typ in ps_types:
+            result.append(ps_types[typ])
+        else:
+            raise KeyError(f"Typestring '{typ}' does not correspond to a known subclass of PipelineStep or Data:\n{ps_types}\n{d_types}")
+    return tuple(result)
